irt analysis

+++ srcs/whisper/addswitching2ratingcsv.py
@@ -3,8 +3,11 @@ from pathlib import Path
 
 ratingcsv_path = '/Users/gioli/projects/stateswitch/data/rec/svf_annotated/'
 
-for csv_file in Path(ratingcsv_path).glob("*.csv"):
+for csv_file in Path(ratingcsv_path).glob("*wordtimestampsrated.csv"):
     df = pd.read_csv(csv_file)
     df["switch_flag"] = 0
     new_csv_file = ratingcsv_path + csv_file.name.replace('desc-wordtimestampsrated.csv', 'desc-wordtimestampswithswitch.csv')
+    if Path(new_csv_file).exists():
+        print(f"File {new_csv_file} already exists. Skipping.")
+        continue
     df.to_csv(new_csv_file, index=False)
